+++ R/optimizers.R
@@ -2,6 +2,7 @@
 #'
 #' List of all the optimizer functions:
 #' \itemize{
+#'   \item Conditional Threading \code{\link{opt_cond_thread}}
 #'   \item Constant Folding \code{\link{opt_constant_folding}}
 #'   \item Constant Propagation \code{\link{opt_constant_propagation}}
 #'   \item Dead Code Elimination \code{\link{opt_dead_code}}
@@ -9,18 +10,21 @@
 #'   \item Dead Expression Elimination \code{\link{opt_dead_expr}}
 #'   \item Common Subexpression Elimination \code{\link{opt_common_subexpr}}
 #'   \item Loop-invariant Code Motion \code{\link{opt_loop_invariant}}
+#'   \item Memory Allocation \code{\link{opt_memory_alloc}}
 #' }
 #'
 #' @export
 #'
 all_optimizers <- list(
+  "Conditional Threading" = opt_cond_thread,
   "Constant Folding" = opt_constant_folding,
   "Constant Propagation" = opt_constant_propagation,
   "Dead Code Elimination" = opt_dead_code,
   "Dead Store Elimination" = opt_dead_store,
   "Dead Expression Elimination" = opt_dead_expr,
   "Common Subexpression Elimination" = opt_common_subexpr,
-  "Loop-invariant Code Motion" = opt_loop_invariant
+  "Loop-invariant Code Motion" = opt_loop_invariant,
+  "Memory Allocation" = opt_memory_alloc
 )
 
 #' Max optimizers list.
@@ -29,6 +33,7 @@ all_optimizers <- list(
 #' enabled.
 #' Note that using this optimizers could change the semantics of the program!
 #' \itemize{
+#'   \item Conditional Threading \code{\link{opt_cond_thread}}
 #'   \item Constant Folding \code{\link{opt_constant_folding}}
 #'   \item Constant Propagation \code{\link{opt_constant_propagation}}
 #'   \item Dead Code Elimination \code{\link{opt_dead_code}}
@@ -36,11 +41,13 @@ all_optimizers <- list(
 #'   \item Dead Expression Elimination \code{\link{opt_dead_expr}}
 #'   \item Common Subexpression Elimination \code{\link{opt_common_subexpr}}
 #'   \item Loop-invariant Code Motion \code{\link{opt_loop_invariant}}
+#'   \item Memory Allocation \code{\link{opt_memory_alloc}}
 #' }
 #'
 #' @export
 #'
 max_optimizers <- list(
+  "Conditional Threading" = opt_cond_thread,
   "Constant Folding" = function(texts) {
     opt_constant_folding(texts, fold_floats = TRUE, in_fun_call = TRUE)
   },
@@ -53,5 +60,6 @@ max_optimizers <- list(
   "Common Subexpression Elimination" = function(texts) {
     opt_common_subexpr(texts, in_fun_call = TRUE)
   },
-  "Loop-invariant Code Motion" = opt_loop_invariant
+  "Loop-invariant Code Motion" = opt_loop_invariant,
+  "Memory Allocation" = opt_memory_alloc
 )