+++ NEWS.md
@@ -1,3 +1,8 @@
+# rco 1.0.2
+
+  - Adding Memory Allocation optimizer.
+  - Adding Conditional Threading optimizer.
+
 # rco 1.0.1
 
   - Bug fixing.